requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
DateTime==4.4
idna==3.3
numpy==1.23.0
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
zope.interface==5.4.0
